Replaced TimeRange with TimeRangeWithBars in getVisibleRange() – #335
TimeRangeWithBars extends the TimeRange interface to include the amount of bars missing before and after the data set in order to fill the entire chart viewport.

The idea is to use these fields to determine the amount of data to load dynamically, as the user scrolls and zooms in the chart.